# specs/001-noorestan-catalog/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Noorestan Catalog API
  version: 1.0.0-draft
  description: >-
    Contract for the public catalog and protected administration application. There are no cart,
    checkout, payment, customer-account, online-order, inventory, or visitor-inquiry endpoints.
servers:
  - url: /api/v1
...
paths:
...
  /public/categories/{slug}/filters:
    get:
      tags: [Public]
      operationId: getCategoryFilters
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: Applicable filter definitions and available values
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FilterDefinition' }
        '404': { $ref: '#/components/responses/NotFound' }
  /public/products:
    get:
      tags: [Public]
      operationId: listPublicProducts
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 200 } }
        - { name: category, in: query, schema: { type: string } }
        - name: filter
          in: query
          description: Repeatable `definitionKey:value` category filter.
          schema: { type: array, items: { type: string }, maxItems: 20 }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: Published products only
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedProducts' }
        '400': { $ref: '#/components/responses/ValidationError' }
  /public/products/{slug}:
    get:
      tags: [Public]
      operationId: getPublicProduct
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: Published product details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-noorestan-session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
  parameters:
...
    Slug: { name: slug, in: path, required: true, schema: { type: string, maxLength: 160 } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 24 } }
...
  responses:
...
    NotFound:
      description: Resource does not exist or is not public
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
    ValidationError:
      description: Request validation failed
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ValidationProblem' } } }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        correlationId: { type: string }
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties: { type: array, items: { type: string } }
...
    Image:
      type: object
      required: [id, altTextFa, variants]
      properties:
        id: { type: string, format: uuid }
        altTextFa: { type: string }
        isPrimary: { type: boolean }
        variants:
          type: array
          items:
            type: object
            required: [url, width, height, mediaType]
            properties:
              url: { type: string, format: uri-reference }
              width: { type: integer }
              height: { type: integer }
              mediaType: { type: string }
    CategorySummary:
      type: object
      required: [id, slug, nameFa, productCount]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        nameFa: { type: string }
        descriptionFa: { type: string }
        productCount: { type: integer, minimum: 1 }
    FilterDefinition:
      type: object
      required: [key, labelFa, type, options]
      properties:
        key: { type: string }
        labelFa: { type: string }
        type: { type: string, enum: [number, boolean, choice] }
        unitFa: { type: string }
        options: { type: array, items: { type: string } }
    ProductSummary:
      type: object
      required: [id, slug, nameFa, category, shortDescriptionFa]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        nameFa: { type: string }
        category: { $ref: '#/components/schemas/CategorySummary' }
        shortDescriptionFa: { type: string }
        primaryImage: { $ref: '#/components/schemas/Image' }
    PagedProducts:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ProductSummary' } }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/ProductSummary'
        - type: object
          required: [descriptionFa, features, specifications, images, contact]
          properties:
            descriptionFa: { type: string }
            technicalNotesFa: { type: string }
            features: { type: array, items: { type: string } }
            specifications: { type: array, items: { $ref: '#/components/schemas/SpecificationDisplay' } }
            images: { type: array, items: { $ref: '#/components/schemas/Image' } }
            contact: { $ref: '#/components/schemas/ContactLinks' }
    SpecificationDisplay:
      type: object
      required: [key, labelFa, displayValueFa]
      properties:
        key: { type: string }
        labelFa: { type: string }
        displayValueFa: { type: string }
        unitFa: { type: string }
    ContactLinks:
      type: object
      required: [phone, whatsapp, email]
      properties:
        phone: { type: string, format: uri }
        whatsapp: { type: string, format: uri }
        email: { type: string, format: uri }
...